Continue to put everything in classmethod for BasSorter.

+++ spikeinterface/sorters/basesorter.py
@@ -16,7 +16,7 @@ from joblib import Parallel, delayed
 
 from spikeinterface.core import load_extractor
 from spikeinterface.core.core_tools import check_json
-from .sorter_tools import SpikeSortingError
+from .utils import SpikeSortingError
 
 
 
@@ -35,49 +35,94 @@ class BaseSorter:
     handle_multi_segment = False
 
     def __init__(self, recording=None, output_folder=None, verbose=False,
-                 delete_output_folder=False):
-                     #~ grouping_property=None, 
+                remove_existing_folder=False, delete_output_folder=False, ):
+        
+        output_folder = self.initialize_folder(recording, output_folder, verbose, remove_existing_folder)
+        
+        self.recording = recording
+        self.verbose = verbose
+        self.delete_output_folder = delete_output_folder
+        self.output_folder = output_folder
+    
+    def set_params(self, sorter_params):
+        """
+        Mimic the old API
+        This should not be used anymore but still works.
+        """
+        p = self.set_params_to_folder(self.recording, self.output_folder, sorter_params, self.verbose)
+        self.params = p
+    
+    def run(self, raise_error=True):
+        """
+        Main function keept for backward compatibility.
+        This should not be used anymore but still works.
+        """
+        # setup recording
+        self.setup_recording(self.recording, self.output_folder, self.params, self.verbose)
+        
+        # compute
+        self.run_from_folder(self.output_folder, raise_error=True)
 
-        assert self.is_installed(), """The sorter {} is not installed.
-        Please install it with:  \n{} """.format(self.sorter_name, self.installation_mesg)
-        if self.requires_locations:
+ 
+    def get_result(self):
+        sorting = self.get_result_from_folder(self.output_folder)
+        if self.delete_output_folder:
+            shutil.rmtree(str(self.output_folder), ignore_errors=True)
+        return sorting
+    
+    
+    #############################################
+    
+    # class method zone
+    
+    @classmethod
+    def initialize_folder(cls, recording, output_folder, verbose, remove_existing_folder):
+        assert cls.is_installed(), """The sorter {} is not installed.
+        Please install it with:  \n{} """.format(cls.sorter_name, cls.installation_mesg)
+        
+        if cls.requires_locations:
             locations = recording.get_channel_locations()
             if locations is None:
                 raise RuntimeError("Channel locations are required for this spike sorter. "
                                    "Locations can be added to the RecordingExtractor by loading a probe file "
                                    "(.prb or .csv) or by setting them manually.")
 
-        self.verbose = verbose
-        #~ self.grouping_property = grouping_property
-        self.params = self.default_params()
+        params = cls.default_params()
 
         if output_folder is None:
-            output_folder = self.sorter_name + '_output'
-        output_folder = Path(output_folder).absolute()
+            output_folder = cls.sorter_name + '_output'
         
+        # .absolute() not anymore
+        output_folder = Path(output_folder) 
 
+        print('remove_existing_folder', remove_existing_folder)
         if output_folder.is_dir():
-            shutil.rmtree(str(output_folder))
-
-        self.output_folder = output_folder
-        self.output_folder.mkdir(parents=True, exist_ok=True)
+            if remove_existing_folder:
+                shutil.rmtree(str(output_folder))
+            else:
+                raise ValueError(f'Folder {output_folder} already exsits')
+        
+    
+        output_folder.mkdir(parents=True, exist_ok=True)
 
         
         if recording.get_num_segments() > 1:
-            if not self.handle_multi_segment:
-                raise ValueError(f'This sorter {self.sorter_name} do not handle multi segment, use recording.split(by=...)')
+            if not cls.handle_multi_segment:
+                raise ValueError(f'This sorter {cls.sorter_name} do not handle multi segment, use recording.split(by=...)')
         
-        self.recording = recording
+        #~ self.recording = recording
         
-        rec_file = self.output_folder / 'spikeinterface_recording.json'
+        rec_file = output_folder / 'spikeinterface_recording.json'
+        #~ assert recording.is_dumpable
+        #~ recording.dump_to_json(rec_file)
         if recording.is_dumpable:
             recording.dump_to_json(rec_file)
         else:
             d = {'warning': 'The recording is not dumpable'}
             rec_file.write_text(json.dumps(d, indent=4), encoding='utf8')
-        
-        self.delete_output_folder = delete_output_folder
 
+        return output_folder
+    
     @classmethod
     def default_params(cls):
         return copy.deepcopy(cls._default_params)
@@ -85,61 +130,53 @@ class BaseSorter:
     @classmethod
     def params_description(cls):
         return copy.deepcopy(cls._params_description)
-        
-    def set_params(self, **params):
+    
+    @classmethod
+    def set_params_to_folder(cls, recording, output_folder, new_params, verbose):
+        params = cls.default_params()
+
+        # verify paarms are in list
         bad_params = []
-        for p in params.keys():
-            if p not in self._default_params.keys():
+        for p in new_params.keys():
+            if p not in params.keys():
                 bad_params.append(p)
         if len(bad_params) > 0:
             raise AttributeError('Bad parameters: ' + str(bad_params))
-        self.params.update(params)
+        
+        params.update(new_params)
+        
+        # custum check params
+        params = cls._check_params(recording, output_folder, params)
+        # common check : filter warning
+        if recording.is_filtered and cls._check_already_filtered(params) and verbose:
+            print(f"Warning! The recording is already filtered, but {cls.sorter_name} filter is enabled")
 
         # dump parameters inside the folder with json
-        self._dump_params()
+        cls._dump_params(recording, output_folder, params, verbose)
         
-        # filter warning
-        if self.recording.is_filtered and self._check_already_filtered(params) and self.verbose:
-            print(f"Warning! The recording is already filtered, but {self.sorter_name} filter is enabled")
-
-    def _dump_params(self):
-        #~ for output_folder, recording in zip(self.output_folders, self.recording_list):
-            #~ with open(str(output_folder / 'spikeinterface_params.json'), 'w', encoding='utf8') as f:
-                #~ params = dict()
-                #~ params['sorter_params'] = self.params
-                #~ # only few properties are put to json
-                #~ params['recording'] = recording.to_dict(include_properties=False, include_features=False)
-                #~ json.dump(check_json(params), f, indent=4)
-
-        with (self.output_folder / 'spikeinterface_params.json').open(mode='w', encoding='utf8') as f:
+        return params
+        
+    @classmethod
+    def _dump_params(cls, recording, output_folder, sorter_params, verbose):
+        with (output_folder / 'spikeinterface_params.json').open(mode='w', encoding='utf8') as f:
             params = dict()
-            params['sorter_name'] = self.sorter_name
-            params['sorter_params'] = self.params
-            params['verbose'] = self.verbose
-            
-            # only few properties/features are put to json
-            #~ params['recording'] = self.recording.to_dict(include_properties=False, include_features=False)
-            #~ from pprint import pprint
-            #~ pprint(params)
+            params['sorter_name'] = cls.sorter_name
+            params['sorter_params'] = sorter_params
+            params['verbose'] = verbose
             json.dump(check_json(params), f, indent=4)
 
-    def run(self, raise_error=True):
-        """
-        Main function keept for backward compatibility.
-        This should not be used anymore.
-        """
-        # setup recording
-        self._setup_recording(self.recording, self.output_folder)
+    
+    @classmethod
+    def setup_recording(cls, recording, output_folder, params, verbose):
+        output_folder = Path(output_folder)
+        cls._setup_recording(recording, output_folder, params, verbose)
         
-        # compute
-        self.compute_from_folder(self.output_folder, raise_error=True)
-
-    @staticmethod
-    def compute_from_folder(output_folder, raise_error=True):
+    @classmethod
+    def run_from_folder(cls, output_folder, raise_error):
         # need setup_recording to be done.
         output_folder = Path(output_folder)
         
-        # retrieve sorter and params
+        # retrieve sorter_name and params
         with (output_folder / 'spikeinterface_params.json').open(mode='r') as f:
             params = json.load(f)
         sorter_params = params['sorter_params']
@@ -161,10 +198,10 @@ class BaseSorter:
         }
         t0 = time.perf_counter()
         
-        #~ SorterClass._compute_from_folder(output_folder, sorter_params, verbose)
+        #~ SorterClass._run_from_folder(output_folder, sorter_params, verbose)
         
         try:
-            SorterClass._compute_from_folder(output_folder, sorter_params, verbose)
+            SorterClass._run_from_folder(output_folder, sorter_params, verbose)
             t1 = time.perf_counter()
             run_time = float(t1 - t0)
             has_error = False
@@ -199,68 +236,74 @@ class BaseSorter:
                 print(f'{sorter_name} run time {run_time:0.2f}s')
 
         if has_error and raise_error:
-            raise SpikeSortingError(f"Spike sorting failed. You can inspect the runtime trace in "
-                                    f"the {sorter_name}.log of the output folder.'")
+            raise SpikeSortingError("Spike sorting failed. You can inspect the runtime trace in spikeinterface_log.json")
         
         return run_time
 
-    def get_result(self):
-        sorting = self.get_result_from_folder(self.output_folder)
+    @classmethod
+    def get_result_from_folder(cls, output_folder):
+        output_folder = Path(output_folder)
+        # check erros in log file
+        log_file  = output_folder / 'spikeinterface_log.json'
+        if not log_file.is_file():
+            raise SpikeSortingError('get reult error: the folder do not contain spikeinterface_log.json')
 
-        if self.delete_output_folder:
-            #~ for out in self.output_folders:
-                #~ if self.verbose:
-                    #~ print("Removing ", str(out))
-                #~ shutil.rmtree(str(out), ignore_errors=True)
-            shutil.rmtree(str(self.output_folder), ignore_errors=True)
+        with log_file.open('r', encoding='utf8') as f:
+            log = json.load(f)
         
-        if sorting.get_sampling_frequency() is None:
-            sorting.set_sampling_frequency(self.recording.get_sampling_frequency())
-
+        if bool(log['error']):
+            raise SpikeSortingError("Spike sorting failed. You can inspect the runtime trace in spikeinterface_log.json")
+        
+        sorting = cls._get_result_from_folder(output_folder)
         return sorting
 
-    #############################################"
+    #############################################
     
     # Zone to be implemeneted
-    def _setup_recording(self, recording, output_folder):
+    # by design all are implemented with class method.
+    # No instance!!
+    # So "self" is not available. Everything is folder based.
+    # This should help for computing distribution
+
+    @classmethod
+    def get_sorter_version(cls):
+        # need be implemented in subclass
+        raise NotImplementedError
+
+    @classmethod
+    def _check_params(cls, recording, output_folder, params):
+        # need be implemented in subclass for custum checks
+        # here nothing to change
+        return params
+
+    @classmethod
+    def _setup_recording(cls, recording, output_folder, params, verbose):
         # need be implemented in subclass
         # this setup ONE recording (or SubExtractor)
         # this must copy (or not) the trace in the appropirate format
         # this must take care of geometry file (PRB, CSV, ...)
         # this must generate all needed script
         raise NotImplementedError
-    
-    @classmethod
-    def get_sorter_version(cls):
-        # need be implemented in subclass
-        raise NotImplementedError
-    
+
     @classmethod
     def is_installed(cls):
         # need be implemented in subclass
         raise NotImplementedError
-    
+
     @classmethod
     def _check_already_filtered(cls, params):
         return False
         # need be implemented in subclass for custum checks
-    
 
-    #~ def _run(self, recording, output_folder):
-        #~ # need be implemented in subclass
-        #~ # this run the sorter on ONE recording (or SubExtractor)
-        #~ # this must run or generate the command line to run the sorter for one recording
-        #~ raise NotImplementedError
-    
     @classmethod
-    def _compute_from_folder(cls, output_folder, params, verbose):
+    def _run_from_folder(cls, output_folder, params, verbose):
         # need be implemented in subclass
         # this is where the script is launch for one recording from a folder already prepared
         # this must run or generate the command line to run the sorter for one recording
         raise NotImplementedError
 
     @classmethod
-    def get_result_from_folder(cls, output_folder):
+    def _get_result_from_folder(cls, output_folder):
         # need be implemented in subclass
         raise NotImplementedError
